+++ CHANGELOG.md
@@ -21,6 +21,8 @@
 - More efficient account initialization by fetching all account Bitcoin xpubs at once
 - Enable search transactions by note, address, or txid
 - Move "Export" (export transactions) to account info page
+
+## v4.48.7
 - ios: fix Pocket user verification button
 - Change block explorer to mempool.space